pyproject.toml
[tool.poetry]
name = "pressure-based-tumor-detection"
version = "0.1.0"
description = "Using Aruco tracking to localize a system to try to localize tumors using pressure values."
authors = ["Aditya Mangalampalli <aditya.mangalampalli@gmail.com>"]
license = "GPL3"

[tool.poetry.dependencies]
python = "^3.8"
black = "^21.7b0"
esptool = "^3.1"
adafruit-ampy = "^1.1.0"
matplotlib = "^3.4.3"
scikit-learn = "^0.24.2"
pre-commit = "^2.14.1"
opencv-python = "^4.5.3"
opencv-contrib-python = "^4.5.3"
plotly = "^5.3.1"
tkdesigner = "^1.0.4"
pyzmq = "^22.3.0"
msgpack = "^1.0.2"
msgpack-numpy = "^0.4.7"
scapy = "^2.4.5"
Pillow = "^8.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
